Added a map
