+++ .env
@@ -9,3 +9,4 @@ BREVO_HOST=smtp-relay.brevo.com
 BREVO_USER=user
 BREVO_PASSWORD=pass
 SUPPORT_FROM=AICulture-Support@myapp.com
+OPENWEATHER_API=key
