+++ package.json
@@ -34,6 +34,7 @@
     "express": "^4.18.2",
     "jotai": "^2.4.2",
     "js-sha256": "^0.10.1",
+    "leaflet": "^1.9.4",
     "next": "^13.2.4",
     "next-auth": "^4.23.1",
     "nodemailer": "^6.9.5",
@@ -42,6 +43,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-file-base64": "^1.0.3",
+    "react-leaflet": "^4.2.1",
     "react-player": "^2.15.1",
     "react-router-dom": "^6.22.1",
     "sharp": "^0.32.6",
@@ -50,6 +52,7 @@
     "typescript": "5.2.2"
   },
   "devDependencies": {
+    "@types/leaflet": "^1.9.12",
     "@types/node": "^20.8.3",
     "@types/nodemailer": "^6.4.14",
     "eslint-config-prettier": "^9.0.0",
