+++ src/app/contact/page.tsx
@@ -7,6 +7,8 @@ import { useEffect, useState } from "react";
 import "../globalicons.css";
 import { NavBar } from "@/components/navbar/navbar";
 import { useSession } from "next-auth/react";
+import { Mice } from "@/components/mice/mouse";
+import { MenuBar } from "@/components/menubar/menubar";
 
 export default function Contact() {
   const [name, setName] = useState("");
@@ -50,38 +52,7 @@ export default function Contact() {
           });
         }}
       >
-        <Image
-          src={"/cursor.png"}
-          alt="cursor"
-          id="mouse"
-          width={30}
-          height={30}
-          style={{
-            display: "block",
-            zIndex: 9999,
-            position: "absolute",
-            pointerEvents: "none",
-          }}
-          onClick={(l) => {
-            return true;
-          }}
-        ></Image>
-        <Image
-          src={"/pointer.png"}
-          alt="cursor"
-          id="pointer"
-          width={20}
-          height={30}
-          style={{
-            display: "none",
-            zIndex: 9999,
-            position: "absolute",
-            pointerEvents: "none",
-          }}
-          onClick={(l) => {
-            return true;
-          }}
-        ></Image>
+        <Mice />
         <nav
           style={{
             display: "flex",
@@ -94,6 +65,7 @@ export default function Contact() {
         >
           <NavBar />
         </nav>
+        <MenuBar />
         <Stack
           style={{
             alignItems: "center",
