+++ src/app/globalicons.css
@@ -132,4 +132,11 @@
   text-align: left;
   background-color: #ffffff; 
  }
- 
+ #arrow {
+  display: flex;
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 9999;
+  cursor: pointer;
+ }
